+++ crates/forge/tests/integration/gas.rs
@@ -1,12 +1,9 @@
-use crate::utils::runner::{Contract, assert_gas, assert_passed, capture_assertion_panic};
+use crate::utils::runner::{Contract, assert_gas, assert_passed};
 use crate::utils::running_tests::run_test_case;
 use crate::utils::test_case;
 use forge_runner::forge_config::ForgeTrackedResource;
-use forge_runner::test_case_summary::{AnyTestCaseSummary, TestCaseSummary};
-use forge_runner::test_target_summary::TestTargetSummary;
 use indoc::{formatdoc, indoc};
 use shared::test_utils::node_url::node_rpc_url;
-use shared::test_utils::output_assert::assert_stdout_contains;
 use starknet_api::execution_resources::{GasAmount, GasVector};
 use std::path::Path;
 
@@ -23,155 +20,6 @@ use std::path::Path;
 // but must remain within a reasonable range
 //
 // Sierra syscall l2 gas: `n_steps * step_gas_cost + sum(builtin_count * builtin_gas_cost)`
-
-#[test]
-fn assert_gas_failure_shows_gas_diff_and_test_case_name() {
-    let test = test_case!(indoc!(
-        r"
-            #[test]
-            fn gas_assertion_diagnostics() {
-                keccak::keccak_u256s_le_inputs(array![1].span());
-            }
-        "
-    ));
-
-    let result = run_test_case(&test, ForgeTrackedResource::CairoSteps);
-
-    assert_passed(&result);
-
-    // Intentionally wrong expectation so that `assert_gas` fails and we can inspect the diagnostics.
-    let panic_message = capture_assertion_panic(|| {
-        assert_gas(
-            &result,
-            "gas_assertion_diagnostics",
-            GasVector {
-                l1_gas: GasAmount(1),
-                l1_data_gas: GasAmount(2),
-                l2_gas: GasAmount(3),
-            },
-        );
-    });
-
-    assert_stdout_contains(
-        panic_message.clone(),
-        indoc! {r"
-        Gas assertion failed for test case `test_package_integrationtest::test_case::gas_assertion_diagnostics`.
-        expected: l1_gas: 1, l1_data_gas: 2, l2_gas: 3
-        actual:   l1_gas: [..], l1_data_gas: [..], l2_gas: [..]
-        diff:     l1_gas: [..], l1_data_gas: [..], l2_gas: [..]
-        "},
-    );
-
-    // `diff` must equal the absolute difference between `expected` and `actual`.
-    let actual = parse_gas_line(&panic_message, "actual:");
-    let diff = parse_gas_line(&panic_message, "diff:");
-    assert_eq!(diff.l1_gas.0, actual.l1_gas.0.abs_diff(1));
-    assert_eq!(diff.l1_data_gas.0, actual.l1_data_gas.0.abs_diff(2));
-    assert_eq!(diff.l2_gas.0, actual.l2_gas.0.abs_diff(3));
-}
-
-#[test]
-fn assert_gas_reports_when_test_case_is_missing() {
-    let summaries = summaries(vec![
-        single_ignored("pkg::module::some_other_test"),
-        single_ignored("pkg::module::another_test"),
-    ]);
-
-    let panic_message = capture_assertion_panic(|| {
-        assert_gas(&summaries, "missing_test", GasVector::default());
-    });
-
-    assert_stdout_contains(
-        panic_message,
-        indoc! {r"
-        Gas assertion failed: test case `missing_test` was not found. Available test cases:
-         - pkg::module::some_other_test
-         - pkg::module::another_test
-        "},
-    );
-}
-
-#[test]
-fn assert_gas_rejects_fuzzing_test_case() {
-    let summaries = summaries(vec![fuzzing_ignored("pkg::module::fuzzed")]);
-
-    let panic_message = capture_assertion_panic(|| {
-        assert_gas(&summaries, "fuzzed", GasVector::default());
-    });
-
-    assert_eq!(panic_message, "Cannot use assert_gas! for fuzzing tests");
-}
-
-#[test]
-fn assert_gas_reports_non_passed_test_case() {
-    let summaries = summaries(vec![single_failed("pkg::module::failing")]);
-
-    let panic_message = capture_assertion_panic(|| {
-        assert_gas(&summaries, "failing", GasVector::default());
-    });
-
-    assert_eq!(
-        panic_message,
-        "Gas assertion failed for test case `pkg::module::failing`: expected passed test case with gas information, but test case was failed"
-    );
-}
-
-/// Parses a `l1_gas: <n>, l1_data_gas: <n>, l2_gas: <n>` line labelled with `label`
-/// out of an `assert_gas` diagnostic message.
-fn parse_gas_line(message: &str, label: &str) -> GasVector {
-    let line = message
-        .lines()
-        .find(|line| line.trim_start().starts_with(label))
-        .unwrap_or_else(|| panic!("no `{label}` line found in message:\n{message}"));
-
-    let value = |key: &str| -> u64 {
-        let after = line
-            .split(&format!("{key}: "))
-            .nth(1)
-            .unwrap_or_else(|| panic!("`{key}` not found in line: {line}"));
-        after
-            .split(|c: char| !c.is_ascii_digit())
-            .next()
-            .filter(|s| !s.is_empty())
-            .unwrap_or_else(|| panic!("no number after `{key}` in line: {line}"))
-            .parse()
-            .unwrap()
-    };
-
-    GasVector {
-        l1_gas: GasAmount(value("l1_gas")),
-        l1_data_gas: GasAmount(value("l1_data_gas")),
-        l2_gas: GasAmount(value("l2_gas")),
-    }
-}
-
-fn summaries(cases: Vec<AnyTestCaseSummary>) -> Vec<TestTargetSummary> {
-    vec![TestTargetSummary {
-        test_case_summaries: cases,
-    }]
-}
-
-fn single_failed(name: &str) -> AnyTestCaseSummary {
-    AnyTestCaseSummary::Single(TestCaseSummary::Failed {
-        name: name.to_string(),
-        msg: None,
-        debugging_trace: None,
-        fuzzer_args: Vec::new(),
-        test_statistics: (),
-    })
-}
-
-fn single_ignored(name: &str) -> AnyTestCaseSummary {
-    AnyTestCaseSummary::Single(TestCaseSummary::Ignored {
-        name: name.to_string(),
-    })
-}
-
-fn fuzzing_ignored(name: &str) -> AnyTestCaseSummary {
-    AnyTestCaseSummary::Fuzzing(TestCaseSummary::Ignored {
-        name: name.to_string(),
-    })
-}
 
 #[test]
 fn declare_cost_is_omitted_cairo_steps() {
